Reuse MIT License from Miguel Grinberg
This commit adds the MIT License originally authored by Miguel Grinberg
to the project. The license grants permission to use, copy, modify,
merge, publish, distribute, sublicense, and/or sell copies of the Software,
subject to certain conditions outlined in the license.

The license text has been included in the LICENSE file in accordance
with the requirements specified in the license.

[MIT License](LICENSE)